requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autopep8==1.4.4
bitfinex-api-py==1.1.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
eventemitter==0.2.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
flask-redis==0.4.0
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-SQLAlchemy-Session==1.1
Flask-WTF==0.14.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.3
pycodestyle==2.5.0
pyee==6.0.0
pylint==2.4.4
python-dotenv==0.10.3
python-engineio==3.11.2
python-socketio==4.4.0
redis==3.3.11
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
urllib3==1.25.7
websockets==8.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
yarl==1.4.2
